+++ sources/TerminalFile.h
@@ -20,7 +20,7 @@ extern NSString *const kTerminalFileShouldStopNotification;
 // You must call -download after initWithName:size: to enter starting status.
 // A nil name opens a save panel.
 // A size of -1 means the size is unknown.
-- (instancetype)initWithName:(NSString *)name size:(int)size;
+- (instancetype)initWithName:(NSString *)name size:(NSInteger)size;
 
 // Appends data to a file in transferring status. Enters transferring status.
 - (void)appendData:(NSString *)data;
