+++ sources/TransferrableFile.h
@@ -25,7 +25,7 @@ typedef NS_ENUM(NSInteger, TransferrableFileStatus) {
 @property(atomic, assign) BOOL openWhenFinished;
 @property(atomic, assign) TransferrableFileStatus status;
 @property(atomic, assign) NSUInteger bytesTransferred;
-@property(atomic, assign) int fileSize;  // -1 if unknown
+@property(atomic, assign) NSInteger fileSize;  // -1 if unknown
 @property(atomic, retain) TransferrableFile *successor;
 @property(atomic, assign) BOOL hasPredecessor;
