+++ sources/VT100TerminalDelegate.h
@@ -305,9 +305,10 @@ typedef NS_ENUM(int, VT100TerminalColorIndex) {
 
 // Download of a base64-encoded file
 // nil = name unknown, -1 = size unknown. Return YES to accept it.
-- (BOOL)terminalWillReceiveFileNamed:(NSString *)name ofSize:(int)size;
+- (BOOL)terminalWillReceiveFileNamed:(NSString *)name
+                              ofSize:(NSInteger)size;
 - (BOOL)terminalWillReceiveInlineFileNamed:(NSString *)name
-                                    ofSize:(int)size
+                                    ofSize:(NSInteger)size
                                      width:(int)width
                                      units:(VT100TerminalUnits)widthUnits
                                     height:(int)height
